# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AdBoard API",
    "version": "v1"
  },
  "paths": {
...
    "/Comment/{id}": {
      "get": {
        "tags": [
          "Comment"
        ],
        "summary": "Получить комментарий по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          },
          "404": {
            "description": "Комментарий с таким id не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Comment"
        ],
        "summary": "Обновляет комментарий по модели",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор комментария",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Модель",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortCommentDto"
                }
              }
            }
          },
          "400": {
            "description": "Модель невалидна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Комментарий не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Конфликт бизнес логики",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Comment"
        ],
        "summary": "Удаляет комментарий по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "204": {
            "description": "Удаление произведено успешно"
          },
          "403": {
            "description": "Нет доступа",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Комментария с таким ID не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/Images/{id}": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Получить изображение по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор изображения",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос выполнен успешно"
          },
          "404": {
            "description": "Изображение с таким id не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Images"
        ],
        "summary": "Удаляет изображение по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор изображения",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "204": {
            "description": "Удаление произведено успешно"
          },
          "403": {
            "description": "Нет доступа",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Изображение с таким ID не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/Images": {
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Создаёт изображение по модели",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "AdvertId",
                  "Content"
                ],
                "type": "object",
                "properties": {
                  "AdvertId": {
                    "type": "string",
                    "description": "Идентификатор объявления",
                    "format": "uuid"
                  },
                  "Content": {
                    "type": "string",
                    "description": "Массив байтов, содержащий изображение",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "AdvertId": {
                  "style": "form"
                },
                "Content": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "201": {
            "description": "Изображение успешно создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortImageDto"
                }
              }
            }
          },
          "400": {
            "description": "Модель данных запроса невалидна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав доступа к изображению",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Произошёл конфликт бизнес логики",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор",
            "format": "uuid",
            "nullable": true
          },
          "text": {
            "type": "string",
            "description": "Содержимое комментария",
            "nullable": true
          },
          "created": {
            "type": "string",
            "description": "Дата создания",
            "format": "date-time"
          },
          "parentId": {
            "type": "string",
            "description": "Родитель (если комментарий является ответом на другой комментарий)",
            "format": "uuid",
            "nullable": true
          },
          "advert": {
            "$ref": "#/components/schemas/ShortAdvertDto"
          },
          "user": {
            "$ref": "#/components/schemas/ShortUserDto"
          }
        },
        "additionalProperties": false,
        "description": "Информация о комментарии"
      },
...
      "ErrorDto": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string",
            "description": "Код ошибки.",
            "nullable": true
          },
          "userMessage": {
            "type": "string",
            "description": "Сообщение для пользователя.",
            "nullable": true
          },
          "internalErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDto"
            },
            "description": "Вложенные ошибки.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Модель ошибки."
      },
...
      "ShortAdvertDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          },
          "location": {
            "type": "string",
            "description": "Геолокация объявления",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "Дата создания.",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean",
            "description": "Признак актуальности объявления."
          },
          "imagesIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Изображения",
            "nullable": true
          },
          "cost": {
            "type": "number",
            "description": "Цена",
            "format": "double"
          }
        },
        "additionalProperties": false,
        "description": "Краткая информация об объявлении"
      },
...
      "ShortCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор",
            "format": "uuid",
            "nullable": true
          },
          "text": {
            "type": "string",
            "description": "Содержимое комментария",
            "nullable": true
          },
          "created": {
            "type": "string",
            "description": "Дата создания",
            "format": "date-time"
          },
          "parentId": {
            "type": "string",
            "description": "Идентификатор родительского комментария",
            "format": "uuid",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/ShortUserDto"
          }
        },
        "additionalProperties": false,
        "description": "Сокращенная информация о комментарии"
      },
      "ShortImageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Краткое описание изображения"
      },
      "ShortUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Имя пользователя",
            "nullable": true
          },
          "login": {
            "type": "string",
            "description": "Логин",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Краткая информация о пользователе"
      },
      "UpdateCommentDto": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "maxLength": 1024,
            "minLength": 0,
            "type": "string",
            "description": "Содержимое комментария"
          }
        },
        "additionalProperties": false,
        "description": "Модель обновления комментария"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.  \n                        Enter 'Bearer' [space] and then your token in the text input below.\n                        Example: 'Bearer secretKey'",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}